# backend/go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Area API server.",
        "title": "Area API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/area/new": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "area"
                ],
                "summary": "Create an area.",
                "parameters": [
                    {
                        "description": "Fill requestr",
                        "name": "CreateAreaRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/area.CreateAreaRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "one or more parameters are missing or invalid"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/area/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "area"
                ],
                "summary": "Delete an area.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Area ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "400": {
                        "description": "one or more parameters are missing or invalid"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/users/me/areas": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get my areas as a user.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/user.GetUserAreasResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "area.CreateAreaRequest": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "object",
                    "properties": {
                        "params": {
                            "type": "object",
                            "additionalProperties": {}
                        },
                        "route": {
                            "type": "string"
                        },
                        "service": {
                            "type": "string"
                        }
                    }
                },
                "name": {
                    "type": "string"
                },
                "reaction": {
                    "type": "object",
                    "properties": {
                        "route": {
                            "type": "string"
                        },
                        "service": {
                            "type": "string"
                        },
                        "target": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "user.GetUserAreasResponse": {
            "type": "object",
            "properties": {
                "action_service": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "reaction_service": {
                    "type": "string"
                },
                "route_action_service": {
                    "type": "string"
                },
                "route_reaction_service": {
                    "type": "string"
                }
            }
        },
...
    }
}